Decoupled spotify profile fetching logic into the SignedInWithSpotify component
