+++ app/page.tsx
@@ -6,9 +6,9 @@ import {
   Auth,
   User,
 } from "firebase/auth";
-import { signIn } from "./auth/GoogleAuthFlow";
+
 import { FirebaseApp, initializeApp } from "firebase/app";
-import { useEffect, useState } from "react";
+import { Suspense, useEffect, useState } from "react";
 import SignedInWithGoogleCard from "./components/SignedInWithGoogleCard";
 import SignedInWithSpotifyCard from "./components/SignedInWithSpotifyCard";
 import { Params } from "next/dist/shared/lib/router/utils/route-matcher";
@@ -29,26 +29,16 @@ export default function Home(urlParams: Params) {
   );
   // Gets auth instance
   const [auth, setAuth] = useState<Auth>(getAuth());
+
   // Reference to firebase user object
   const [firebaseUser, setFirebaseUser] = useState<User>();
 
-  // The user profile returned from spotify api
-  const [spotifyUserProfile, setSpotifyUserProfile] = useState<
-    SpotifyUserProfile | false
-  >(false);
   // This function is passed to the signincard to give it access to our user state
   const updateUserFirebase = (newUser: User) => {
     setFirebaseUser(newUser);
   };
 
-  // This function is passed to the signed in with spotify card to give access to our spotify user profile state
-  const updateSpotifyUserProfile = (userProfile: SpotifyUserProfile) => {
-    setSpotifyUserProfile(userProfile);
-  };
-
   useEffect(() => {
-    console.log("Effect ran");
-
     let params: any = undefined;
     if (urlParams.searchParams.data) {
       params = JSON.parse(urlParams.searchParams.data);
@@ -68,16 +58,6 @@ export default function Home(urlParams: Params) {
       localStorage.setItem("accessToken", JSON.stringify(params));
       document.location = "http://localhost:3000";
     }
-
-    // Check if we need to fetch spotify user profile
-    if (localStorage.getItem("accessToken")) {
-      const accessToken = JSON.parse(
-        localStorage.getItem("accessToken")!
-      ).access_token;
-      fetchProfile(accessToken).then((userProfile) =>
-        setSpotifyUserProfile(userProfile)
-      );
-    }
   }, [auth, urlParams]);
 
   return (
@@ -88,10 +68,7 @@ export default function Home(urlParams: Params) {
         photoURL={firebaseUser?.photoURL}
         updateUser={updateUserFirebase}
       />
-      <SignedInWithSpotifyCard
-        spotifyUserProfile={spotifyUserProfile}
-        updateSpotifyUserProfile={updateSpotifyUserProfile}
-      />
+      <SignedInWithSpotifyCard />
     </div>
   );
 }
