+++ package.json
@@ -22,6 +22,7 @@
     "firebase": "^10.0.0",
     "js-cookie": "^3.0.5",
     "next": "13.4.9",
+    "node-fetch": "^3.3.1",
     "postcss": "8.4.25",
     "react": "18.2.0",
     "react-dom": "18.2.0",
@@ -30,6 +31,7 @@
   },
   "devDependencies": {
     "@types/js-cookie": "^3.0.3",
-    "cypress": "^12.17.1"
+    "cypress": "^12.17.1",
+    "encoding": "^0.1.13"
   }
 }
